GUI for Customer Home Page Added
Main thing done was the GUI for the Customer Home Page was added, so everyone has access. The database is also within the application. A few nuget packages were added (like dapper and sqlite). I am hoping that it adds to GitHub correctly, but I will check that. No functionality was added yet. There is a separate class library where we can put the actual classes and functionality of the classes, so it is separate from the UI if we want. That is how the guy in the youtube video did it.