Log In Page GUI added
Log In Page GUI added. No functionality for either the log in or home page. I messed around with a sql script, but it doesn't connect to anything. If you build and run yours, it will probably throw you to the log in page first because that is what program.cs does. You can change that when you are testing. Also, I locked my GUIs after copying one of them, so the size of the screen that appears is the same with text/label boxes about the same size. The text is Rockwell at the moment, but we can change anything as we go of course.