Comments, Customer Class, Encrypt Password, and Get Random UserID Added
I added the functionality for encrypting a password and generating a random user ID. I need to add the remove that randomly chosen user ID from the table still. I added the Customer Class also, but methods were not added to the class yet. I added comments too. Lastly, I created a few different class files. One is for anything the system does, like encrypt the password (or create a customer maybe?). Another one is for anything that touches the database. Any others are for the specific classes we have. I am not sure if that is how you guys want to set up the application. That was just what I was rolling with at the moment. Also, the classes are all done in the ClassLibrary file and everything to do with the GUI and GUI actions is done on the Air3550 side with new forms generated for each screen. Obviously if anything doesn't make sense or needs to change, let me know. I don't have much oop experience, so I am flexible and open to comments.