Create A Customer Functionality Added
Create a customer functionality was added. I have not deleted the user id because we are still testing things. I will add it later. Also, the database that is created from running the application is located in the bin debug folder if you debug or the bin release folder if you don't debug. Lastly, if you click "Don't Have Account?", the new account form appears. There is built in error checking with it already. Pop ups appear when invalid appears (first name, last name, and street are just any string). State and age are dropdowns which was easy to do (combo box with info added in visual studio not in database). The email is checked to see if the user is already in the system. Then if everything is good, a pop up appears that requires "ok" to be pushed, which closes the create account form and pop up, returning the user to the home page. I did create a new customer object when the account is created. That is the only place I used an object. I put create account in the sqlitedataaccess file because the customer doesn't really do that and it touches the database, so I wanted to keep the database sqlite scripts separate from other things. Let me know if anything needs changing.