Fixed Closing Page Bug
The Log In page was the main page generated when the application would start, so I couldn't close it without closing the application, but I also couldn't end the application if I hid the page. I changed the main page depending on whether the user is logged in or not. This fixes it for the logging part, but when a customer switches from the home page to a book flight page and exits, the application will not end. It is a bug that will need to be fixed somehow.