Added Log In Page Functionality
You can log in with the incorrect or correct credentials and it throws an error or logs you in. Enter button does not work and forgot to add hide password functionality. Currently the log in page closes when the customer home page or the create customer page opens, then create customer page closes when the home page is open.