Asterisks added for password input and fixed flow of forms again
Flow of forms is going to be something we have to look at a lot. My solution was to mark the form as IsAccessible and check if it was IsDisposed. This seemed to work, but I just realized that the main program.cs is going to have to be user neutral, so a customer or the marketing manager could use it. That might be handled through the log in button rather than the program.cs button though. We will see. It works with my forms so far. Adding asterisks was easy, so if we want to use that anywhere else, it's not very difficult. I'll probably use it for the view account information page.