Added All Forms for GUIs
I did not actually do the GUIs for most of the pages. I just wanted to make sure all of the forms are currently the same size. I added a few date pickers, but the rest is empty. Obviously we can make them look similar later, but maybe try as you go to make them similar, so there is less work later if that makes sense. I think all of the forms for every page we have is added now though.