Comments Added. Just a summary update since I committed a lot of times.
Account Information is fully functional when it is a new customer. I still need to test and break it though, so I will update if I find any bugs. I still need to make it functional for returning customer. The red X is functional to close out the entire program for the Account Information page only. I need to add that everywhere else. I also changed the program.cs to only have one main page. It is just hidden and called again if/when needbe.