Account Information Functionality is Done
User can log in as new or existing customer. There information is then sent as an object to the home page. The Account Information Button goes to the Edit page. Everything is shown but the password because you can't unhash a hash apparently, but they can change the password if they want. Format is checked. Save button updates database with/without password and produces a pop-up to say changes were successful. Back button has a pop-up to ask the user if they are sure and if they saved any changes, then they are redirected to the home page. Log out button produces a pop up to ask if they are sure and if they saved any changes, then closes all forms and takes the user back to log in which they can log in again or create an account. Lastly, the red x has functionality as well. It will ask if they want to close the application and if they saved their changes, then closes and ends the program.